+++ lib/src/styles/styles.dart
@@ -7,4 +7,5 @@ export 'chat_input_style.dart';
 export 'file_attachment_style.dart';
 export 'llm_chat_view_style.dart';
 export 'llm_message_style.dart';
+export 'suggestion_style.dart';
 export 'user_message_style.dart';